#19 - cleaned up code, iterator has been replaced with streams

+++ src/main/java/com/car/dealer/model/Loan.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 
 @Data
 public class Loan {
-    private BigDecimal year;
+    private Integer year;
     private BigDecimal percent;
     private BigDecimal yearCostWithoutPercent;
     private BigDecimal yearPercentPrice;
